Refactor to add `app/routes` (close #975)
