+++ docs/internals/users.md
@@ -337,7 +337,7 @@ The [User Model](#user-model) includes additional attributes `username` and `slu
 
 > ⚠️ This is an experimental feature, which will likely evolve in the future (possibly towards the implementation of user groups/roles).
 
-When a `Task` is [created](https://fractal-analytics-platform.github.io/fractal-server/reference/fractal_server/app/api/v1/task/#fractal_server.app.api.v1.task.create_task), the attribute `Task.owner` is set equal to `username` or, if not present, to `slurm_user` (there must be at least one to create a Task). With a similar logic, we consider a user to be the _owner_ of a Task if `username==Task.owner` or, if `username` is `None`, we check that `slurm_user==Task.owner`.
+When a `Task` is [created](https://fractal-analytics-platform.github.io/fractal-server/reference/fractal_server/app/api/v1/task/#fractal_server.app.routes.api.v1.task.create_task), the attribute `Task.owner` is set equal to `username` or, if not present, to `slurm_user` (there must be at least one to create a Task). With a similar logic, we consider a user to be the _owner_ of a Task if `username==Task.owner` or, if `username` is `None`, we check that `slurm_user==Task.owner`.
 The following endpoints require a non-superuser user to be the owner of the Task:
 
 - PATCH `/api/v1/task/{task_id}`,
