--- fractal_server/app/api/v1/__init__.py
@@ -1,3 +1,0 @@
-"""
-`api/v1` module
-"""
