+++ fractal_server/app/routes/api/__init__.py
@@ -1,0 +1,19 @@
+"""
+`api` module
+"""
+from fastapi import APIRouter
+
+from ....config import get_settings
+from ....syringe import Inject
+
+
+router_api = APIRouter()
+
+
+@router_api.get("/alive/")
+async def alive():
+    settings = Inject(get_settings)
+    return dict(
+        alive=True,
+        version=settings.PROJECT_VERSION,
+    )
