+++ fractal_server/app/routes/api/v1/task.py
@@ -9,16 +9,16 @@ from fastapi import Response
 from fastapi import status
 from sqlmodel import select
 
-from ....logger import set_logger
-from ...db import AsyncSession
-from ...db import get_db
-from ...models import Task
-from ...models import WorkflowTask
-from ...schemas import TaskCreate
-from ...schemas import TaskRead
-from ...schemas import TaskUpdate
-from ...security import current_active_user
-from ...security import User
+from .....logger import set_logger
+from ....db import AsyncSession
+from ....db import get_db
+from ....models import Task
+from ....models import WorkflowTask
+from ....schemas import TaskCreate
+from ....schemas import TaskRead
+from ....schemas import TaskUpdate
+from ....security import current_active_user
+from ....security import User
 from ._aux_functions import _get_task_check_owner
 
 router = APIRouter()
