+++ fractal_server/app/routes/api/v1/task_collection.py
@@ -13,32 +13,32 @@ from fastapi import status
 from pydantic.error_wrappers import ValidationError
 from sqlmodel import select
 
-from ....config import get_settings
-from ....logger import close_logger
-from ....logger import set_logger
-from ....syringe import Inject
-from ....tasks.collection import _TaskCollectPip
-from ....tasks.collection import create_package_dir_pip
-from ....tasks.collection import create_package_environment_pip
-from ....tasks.collection import download_package
-from ....tasks.collection import get_collection_data
-from ....tasks.collection import get_collection_log
-from ....tasks.collection import get_collection_path
-from ....tasks.collection import get_log_path
-from ....tasks.collection import inspect_package
-from ....tasks.collection import slugify_task_name
-from ...db import AsyncSession
-from ...db import DBSyncSession
-from ...db import get_db
-from ...db import get_sync_db
-from ...models import State
-from ...models import Task
-from ...schemas import StateRead
-from ...schemas import TaskCollectPip
-from ...schemas import TaskCollectStatus
-from ...schemas import TaskCreate
-from ...security import current_active_user
-from ...security import User
+from .....config import get_settings
+from .....logger import close_logger
+from .....logger import set_logger
+from .....syringe import Inject
+from .....tasks.collection import _TaskCollectPip
+from .....tasks.collection import create_package_dir_pip
+from .....tasks.collection import create_package_environment_pip
+from .....tasks.collection import download_package
+from .....tasks.collection import get_collection_data
+from .....tasks.collection import get_collection_log
+from .....tasks.collection import get_collection_path
+from .....tasks.collection import get_log_path
+from .....tasks.collection import inspect_package
+from .....tasks.collection import slugify_task_name
+from ....db import AsyncSession
+from ....db import DBSyncSession
+from ....db import get_db
+from ....db import get_sync_db
+from ....models import State
+from ....models import Task
+from ....schemas import StateRead
+from ....schemas import TaskCollectPip
+from ....schemas import TaskCollectStatus
+from ....schemas import TaskCreate
+from ....security import current_active_user
+from ....security import User
 
 router = APIRouter()
