+++ fractal_server/app/runner/_common.py
@@ -578,7 +578,7 @@ def execute_tasks(
 
     **Note:** At the end of each task, write current metadata to `working_dir /
     METADATA_FILENAME`, so that they can be read as part of the [`get_job`
-    endpoint](../../api/v1/job/#fractal_server.app.api.v1.job.get_job).
+    endpoint](../../api/v1/job/#fractal_server.app.routes.api.v1.job.get_job).
 
     Arguments:
         executor:
