+++ tests/test_task_collection_api.py
@@ -10,12 +10,16 @@ from devtools import debug
 from sqlmodel import select
 
 from .fixtures_tasks import execute_command
-from fractal_server.app.api.v1.task_collection import _background_collect_pip
-from fractal_server.app.api.v1.task_collection import _TaskCollectPip
-from fractal_server.app.api.v1.task_collection import create_package_dir_pip
-from fractal_server.app.api.v1.task_collection import TaskCollectStatus
 from fractal_server.app.models import State
 from fractal_server.app.models import Task
+from fractal_server.app.routes.api.v1.task_collection import (
+    _background_collect_pip,
+)
+from fractal_server.app.routes.api.v1.task_collection import _TaskCollectPip
+from fractal_server.app.routes.api.v1.task_collection import (
+    create_package_dir_pip,
+)
+from fractal_server.app.routes.api.v1.task_collection import TaskCollectStatus
 from fractal_server.config import get_settings
 from fractal_server.syringe import Inject
 from fractal_server.tasks.collection import get_collection_path
